+++ src/ir_NEC.hpp
@@ -71,10 +71,10 @@
  */
 // http://www.hifi-remote.com/wiki/index.php/NEC
 // https://www.sbprojects.net/knowledge/ir/nec.php
-// NEC: LSB first, <start bit><address:16> (or <address:8><inverted address:8>) <command:8><inverted command:8><stop bit>.
-// ONKYO like NEC but 16 independent address and command bits: <start bit><address:16><command:16><stop bit>
+// NEC: LSB first, <start bit><address:16> (or <address:8><inverted address:8>) <command:16><command:8><inverted command:8><stop bit>.
+// ONKYO: like NEC but force to 16 independent address and 16 bit command bits: <start bit><address:16><command:16><stop bit>
 // Standard NEC sends a special fixed repeat frame.
-// NEC2 sends the same full frame after the 110 ms. I have a DVD remote with NEC2.
+// NEC2: like NEC, but for repeat, the same full frame is sent after the 110 ms. I have a DVD remote with NEC2.
 // NEC and NEC 2 only differ in the repeat frames, so the protocol can only be detected correctly after the first repeat.
 // PIONEER (not implemented) is NEC2 with 40 kHz
 //
@@ -150,6 +150,13 @@ void sendNECSpecialRepeat() {
     IrSender.mark(NEC_BIT_MARK);             // + 560
 }
 
+/**
+ * Convert 16 bit address and 16 bit command to 32 bit NECRaw data
+ * If we get a command < 0x100, we send command and then ~command
+ * If we get an address < 0x100, we send 8 bit address and then ~address
+ * !!! Be aware, that this is flexible, but makes it impossible to send e.g. 0x0042 as 16 bit value!!!
+ * To force send 16 bit address, use: sendOnkyo().
+ */
 uint32_t IRsend::computeNECRawDataAndChecksum(uint16_t aAddress, uint16_t aCommand) {
     LongUnion tRawData;
 
@@ -173,21 +180,11 @@ uint32_t IRsend::computeNECRawDataAndChecksum(uint16_t aAddress, uint16_t aComma
  * There is NO delay after the last sent repeat!
  * @param aNumberOfRepeats  If < 0 then only a special NEC repeat frame will be sent by calling NECProtocolConstants.SpecialSendRepeatFunction().
  */
-void IRsend::sendNEC(uint16_t aAddress, uint8_t aCommand, int_fast8_t aNumberOfRepeats) {
+void IRsend::sendNEC(uint16_t aAddress, uint16_t aCommand, int_fast8_t aNumberOfRepeats) {
     sendPulseDistanceWidth(&NECProtocolConstants, computeNECRawDataAndChecksum(aAddress, aCommand), NEC_BITS, aNumberOfRepeats);
 }
 
-/*
- * NEC2 Send frame and repeat the frame for each requested repeat
- * There is NO delay after the last sent repeat!
- * @param aNumberOfRepeats  If < 0 then nothing is sent.
- */
-void IRsend::sendNEC2(uint16_t aAddress, uint8_t aCommand, int_fast8_t aNumberOfRepeats) {
-    sendPulseDistanceWidth(&NEC2ProtocolConstants, computeNECRawDataAndChecksum(aAddress, aCommand), NEC_BITS, aNumberOfRepeats);
-}
-
-/*
- * Repeat commands should be sent in a 110 ms raster.
+/**
  * There is NO delay after the last sent repeat!
  * @param aNumberOfRepeats  If < 0 then only a special repeat frame without leading and trailing space
  *                          will be sent by calling NECProtocolConstants.SpecialSendRepeatFunction().
@@ -196,8 +193,17 @@ void IRsend::sendOnkyo(uint16_t aAddress, uint16_t aCommand, int_fast8_t aNumber
     sendPulseDistanceWidth(&NECProtocolConstants, (uint32_t) aCommand << 16 | aAddress, NEC_BITS, aNumberOfRepeats);
 }
 
-/*
- * Repeat commands should be sent in a 110 ms raster.
+/**
+ * NEC2 Send frame !!! and repeat the frame for each requested repeat !!!
+ * There is NO delay after the last sent repeat!
+ * @param aNumberOfRepeats  If < 0 then nothing is sent.
+ */
+void IRsend::sendNEC2(uint16_t aAddress, uint16_t aCommand, int_fast8_t aNumberOfRepeats) {
+    sendPulseDistanceWidth(&NEC2ProtocolConstants, computeNECRawDataAndChecksum(aAddress, aCommand), NEC_BITS, aNumberOfRepeats);
+}
+
+/**
+ * Apple: Send NEC with fixed 16 bit Apple address 0x87EE.
  * There is NO delay after the last sent repeat!
  * https://en.wikipedia.org/wiki/Apple_Remote
  * https://gist.github.com/darconeous/4437f79a34e3b6441628
@@ -217,7 +223,7 @@ void IRsend::sendApple(uint8_t aDeviceId, uint8_t aCommand, int_fast8_t aNumberO
     sendPulseDistanceWidth(&NECProtocolConstants, tRawData.ULong, NEC_BITS, aNumberOfRepeats);
 }
 
-/*
+/**
  * Sends NEC protocol
  * @param aNumberOfRepeats  If < 0 then only a special repeat frame without leading and trailing space
  *                          will be sent by calling NECProtocolConstants.SpecialSendRepeatFunction().